Fix orphaned hint PostList bugs
In certain cases under OP_NEAR, OP_PHRASE and or-like query operators a
PostList could be deleted while set as the QueryOptimiser "hint"
PostList.  We now have the QueryOptimiser take ownership in these cases
too.

Testcase multiargop1 will be added as part of an imminent change and
will serve as a regression test for this.