Close link upon exception closing remote prog DB
Previously the remote connection stayed open in this situation, and
then when we subsequently wait for the server to exit the connection has
to time out.

We don't want to just skip the wait for the child to exit, since
then the remote server will stay around in the background, and in
the testsuite we could end up with many such servers running.

(cherry picked from commit be346ae6067b983dcc35c90fc6f6e82460eba050)