Track if there are pending changes to writable remote DB
If there aren't any then we now do nothing for commit() or cancel()
calls.  In particular this saves a message exchange when the
WritableDatabase destructor is called when changes have already been
committed with an explicit call to commit() (which is what we recommend
doing, since with an explicit call to commit() you get to see any
exception which gets thrown).