Fix issue with commit() getting stuck after exception
If commit() failed with an exception while trying to add pending changes
(e.g. InvalidArgumentError due to a long term containing zero bytes)
then a subsequent commit() on the same object would fail in the same
way.

Now we clear the pending changes if commit() fails with an exception in
this case (like we already did for failure at other stages in the
commit).