XFAIL new part of termtoolong1 for chert
It's harder to fix this in chert than it was for glass as for chert
flush_postlist_changes() is called from const methods so we can't
just make it non-const so it can call non-const cancel().

It looks like this has been broken for a long time (before chert
even) and nobody's ever complained - I only encountered the issue myself
while trying to debug a different problem.  It's a corner case situation
in a backend we recommend people migrate away from, and the fix would
be excessively invasive so it seems better to just leave this as it is
for 1.4.x (and chert is gone on git master).