Fix segfault due to hint postlist getting released
The QueryOptimiser keeps a pointer to the most recent postlist to
use as a hint.  We already have a mechanism to take ownership of the
hint when another part of the code wants to delete that postlist,
but this had a flaw if the hint postlist was a subpostlist of a
deleted postlist.  This case is pretty rare, but would typically result
in a segfault.

We now invalidate the hint when a postlist is deleted unless (a) the
postlist being deleted is the current hint (in which case we take
ownership as before) or (b) the hint is already owned by the
QueryOptimiser (in which case we know it isn't referenced elsewhere).
This is safe, but somewhat conservative.