lua: Simplify luaStemImplementation::get_description()
The Lua bindings allows passing a Lua function for where the C++
API takes a Xapian::StemImplementation object.  Previously this
function served to implement the stemmer and provide its description,
with nil being passed in the latter case.  This seems more annoying
than helpful as it forces you to implement a description (or else
your code will fail in a logging build of Xapian even if you
never explicitly request the description in your own code), so
now a static description of "luaStemImplementation()" is returned.

This change should allow any Lua stemmer implementations to continue
to work - the code in them to return a description will just no
longer get used.