lua: Fix 32-bit Xapian::BAD_VALUENO wrapping
On platforms where (sizeof(long) == 4), SWIG by default wraps
Xapian::BAD_VALUENO as a negative constant in Lua, which is then
rejected by a check which disallows passing negative values for unsigned
C++ types.

We now direct SWIG to handle Xapian::valueno as double (which is
what all numbers in Lua actually are) which gives us an unsigned
constant, and also eliminates the negative value check.