Fix query expansion on sharded database
The mechanism for passing in which shard a TermList is from wasn't
hooked up and as a result we always think it's from the first shard.
This means our suggested terms may not be as good as they should be
in this situation.

Probably broken by 145503bbe4a5bf702cd13cb2e592111e8d7ca89a but I
haven't tested revisions before/after that to confirm this.