Fix query expansion on sharded database
The mechanism for passing in which shard a TermList is from wasn't
hooked up and as a result we always think it's from the first shard.
This means our suggested terms may not be as good as they should be
in this situation.

(cherry picked from commit 7ee01561fc687f77fce621ef07fabbc064970169)