Rework the histogram edit distance optimisation
Use a bitmap in an unsigned long long instead of an array.  This catches
nearly as many cases (it's only less effective if there are repeated
characters in one or both of the two words being compared (or rather
repeated characters modulo 64, assuming an 8-bit unsigned long long),
but is faster to check, and overall results in a speed up even though we
have to perform the full edit distance calculation a little more often.