Fix path qualfication for omindex workers
The code to add a path to a worker module was broken - it ended up
setting a char* to point to c_str() of a string which had been
deleted.  However a reversed conditional meant that this code was
never actually executed, but that no path qualification happened,
which means that work modules didn't get found with an installed
version of omega.

This change fixes all that and hook up "make installcheck" for
omindextest, which provides an easy way to exercise this case.