Clarify API docs for MSet::get_termfreq()
Make it clear that this considers all documents in the database, not
only those that matched the searched (it would sometimes be useful to be
able to report the number of occurrences of a term in the matched
documents, but it's not something we currently keep track of).

Reported by Tadeusz Sośnierz and Peter Salomonsen.

(cherry picked from commit 092d91523991696152c6b48ce546c4b3b55ac9f3)