No-op GlassTable::readahead_key for overlong keys
Currently, a query may abort with an InvalidArgument error if one of
the terms exceeds the maximum allowed length for glass btree keys.
However, this error is thrown only if the query triggers a readahead_key
call on the glass table.

This patch updates readahead_key to report overlong keys as not found,
rather than aborting the whole query.

Closes https://github.com/xapian/xapian/pull/313