Remove DummyFreqSource API class
This is part of the clustering API.  It's really a left-over from early
development on the clustering feature.  It returns a frequency of 1
whatever term is asked about, and treating all terms identically isn't
going to give useful clusters.

Until my previous commit you couldn't actually pass an object of this
class to anything so now seems a good time to remove it.