Add QueryParser::FLAG_ACCUMULATE
By default, the unstem and stoplist data is reset by a call to
parse_query(), which makes sense if you use the same QueryParser
object to parse a series of independent queries.

If you're using the same QueryParser object to parse several
fields on the same query form, you may want to have the unstem
and stoplist data combined for all of them, in which case you
can use this flag to prevent this data from being reset.