Fix latent bug in remote msg_positionlistcount()
We weren't checking that we'd reached the term we asked for after
skip_to(), which potentially means returning a non-zero answer instead
of zero when the term isn't present.  I don't believe this case can
currently happen, since this call only happens from a RemoteTermList
and the term passed is that at the current position, but we don't
want latent bugs lying in wait for future changes.