Don't specify explicit filename to Doxygen @file
If there's no filename, Doxygen assumes the current file, which is
exactly what we always want.  Listing the filename explicitly just
creates a maintenance task as one has to remember to change it when
a file is renamed or copied.

I've verified the generated API documentation is identical before
and after this change.