Avoid bogus deprecation warning without optimisation
If compiling with GCC without optimisation, GCC emits a deprecation
warning for code in the definition
QueryParser::add_valuerangeprocessor() which is provided for backwards
API compatibility even if this method is never used anywhere.

This isn't helpful, especially if the user is using -Werror, so disable
the -Wdeprecated-deprecations warning for this code.

Issue specific to this branch.  Reported by starmad on #xapian.