Stop converting OP_NEAR/OP_PHRASE to OP_AND
Previously we did this conversion if there was no positional
information in the database, but it's a potentially surprising
edge case behaviour, and surprisingly complicated to handle
internally.

The intention is to replace this with a new QueryParser flag
which results in a Query which doesn't use positional information.
That's more flexible as it can be enabled even if the database
does have positions.