Avoid checking for permissions on every query
Only queries which use positional information now check, which
should help improve notmuch delete performance.  Thanks to andreas
on #notmuch for analysis of the problem.

This change isn't relevant to git master, since there we've dropped
the feature of changing OP_PHRASE and OP_NEAR to OP_AND if there
are no positions in the database being searched.