Add QueryParser::FLAG_NO_POSITIONS flag
With this flag enabled, no positional information will be used and any
query operations which would use it are replaced by the nearest
equivalent which doesn't (so phrase searches, NEAR and ADJ will result
in OP_AND).  This is intended to replace the automatic conversion of
OP_PHRASE, etc to OP_AND when a database has no positional information,
which we stopped doing recently on git master.