Always restore compound_weight flag after OP_SYNONYM
Previously the flag was left set if the OP_SYNONYM optimised away
to MatchNothing.  Following the code flow, this doesn't actually
seem to be able to cause incorrect behaviour, but it may cause us to
think we can't apply some further optimisations to parts of the query
processed after such an OP_SYNONYM.  Also we want to be robust to
future code changes.