Fix topterms template to not trigger early matching
This template was checking $msize before including the `query`
template, but doing so would trigger the query to be run, which
means that settings early in the `query` template which would
affect the result (such as $setmap{prefix,...}) were being
ignored when the `topterms` template was used.

Partly addresses #815, reported by Gennadiy.