Clean up handling of libs we link libxapian with
Previously any libraries explicitly specified to configure by the
user via LIBS=... and also -lm, if configure determined it is needed,
could get added to XAPIAN_LIBS multiple times, as well as also getting
added to the libxapian link command anyway as per automake/libtool
standard handling.

Now they'll only get added once, via the default handling.

Specifying a library more than once on the link line is not a problem
on common platforms, but may be an issue somewhere (and it's on less
common platforms where the user might have to specify LIBS to configure
and/or where -lm may be needed).