Workaround NFS issue in testsuite rm_rf()
This function is used to remove databases by the testharness.

On NFS, rmdir() can fail with EEXIST or ENOTEMPTY (POSIX allows either)
due to .nfs* files which are used by NFS clients to implement the Unix
semantics of a deleted but open file continuing to exist.  We now sleep
and retry a few times in this situation to give the NFS client a chance
to process the closing of the open handle.

Issue mentioned in #631.

(cherry picked from commit 8ef5a8c5f310dc7a74349610fa626acde256718f)