ruby: Fix build with Ruby 3.0 and clang
When building with clang, the Ruby 3.0 headers try to use __declspec
and compilation fails on non-Microsoft platforms unless you pass a
special compiler option.  We now suppress this so that the Ruby
headers use the __attribute__ syntax instead.

Reported by Eduard Panov on macos, where clang is the default compiler.