Drop undocumented HTML formatting encoding feature
1.3.2 introduced a feature where some HTML formatting was preserved
in the generated sample, encoded using different whitespace characters.
However, this feature was never documented (nor even mentioned in the
NEWS file), only applied to documents in HTML or processed via HTML,
and nothing actually uses this formatting information.

It can also be confusing when looking at an Omega database with
xapian-delve, since the document data contains carriage returns.

I think the original intent was that the sample generating/selecting
code might use this information, but in hindsight I'm not sure that
it's very useful there even.