Fix TermIterator::skip_to() sharded DB bug
The OrTermList::skip_to() method which handles this internally was
advancing the left child or right child or both based on whether the
current positions are <, > or ==, but really it should advance both
children whatever the current position order is.  This means that the
current code will sometimes fail to skip all the way to the requested
term.

Uncovered while addressing warning from GCC's -Wduplicated-cond,
reported by dcb in #816.