[php] Reject PHP8 in configure (for now)
SWIG 4.1.0 will supports PHP8, but (a) not released yet and (b) it
changes to wrapping classes via PHP's C API rather than generating
a .php script with the class definitions.  That's an improvement,
but means we'll need to make some adjustments as we're currently
post-processing the generated xapian.php to adjust some things, and
there will no longer be a generated xapian.php.