omindex_poppler: Remove whitespace cleaning pass
There's only a reason to clean up whitespace here because it would save
having to send it over the pipe and would save TermIndexer::index_text()
from having to process the redundant whitespace.  Any stored sample gets
whitespace cleaned anyway.

If there is no redundant whitespace then it is pure overhead to try to
deal with it here, and if there isn't much then the cost of dealing with
it here likely outweighs the saving, and testing on a variety of PDF
files I have to hand I can't find a single one where clear_text()
actually changes the passed text!