Improve ValueRangePostList::estimate_termfreqs()
Estimate by scaling proportional to the termfreq estimate which we
calculate based on the range ends and slot value bounds.

This requires an extra member variable, and on reflection I'm not sure
this is worth improving - this method is only used below an OP_SYNONYM
and that seems an odd place to use ValueRangePostList.