Optimise termfreq handling in alldocs optimisation
Under certain conditions, we use an all-docs PostList for a term
which indexes all documents.  The check for such a term now always
creates the actual PostList first and gets the termfreq from that
instead of asking the database for the termfreq.  This reduces the
work in the common case where the term doesn't match all documents
and is similar work in the case where it does (for glass, there's
the extra overhead of creating a cursor but we still need to read and
decode the same data).