Improve error if block overwritten in writable DB
Drop "are there multiple writers?" as it's rarely a useful question to
ask since we started using fcntl() locking as it's now very hard to
get multiple concurrent writers on a database.  Instead suggest running
xapian-check, which is probably the best next step for a user who hits
this problem.