scriptindex: Allow control of missing unique action
Previously if the index script contained a unique action, a record
which didn't trigger it would be added as a new document.  That's
may be what's wanted sometimes, but in others it might hide an
error in the input data so provide an optional parameter to allow
the behaviour here to be specified, and default to adding a new
document but issuing a warning.

(cherry picked from commit 89bc91b1f6031a7a2bd8ca42e3bc39c0c77d4c59)