scriptindex: Allow control of missing unique action
Previously if the index script contained a unique action, a record
which didn't trigger it would be added as a new document.  That's
may be what's wanted sometimes, but in others it might hide an
error in the input data so provide an optional parameter to allow
the behaviour here to be specified, and default to exiting with an
error.