omindex: Improve handling of worker errors
Workers now exit with EX_OSERR if we can't run the helper,
EX_UNAVAILABLE if we can't initialise the library (or similar problems)
and EX_TEMPFAIL if the timeout is reached.

The first two are treated as hard errors and the same worker won't be
tried again for the rest of the run.  EX_TEMPFAIL and other exit codes
and dying with a signal are assumed to be specific to the input file.

The error message is now much more specific, with a fallback of
reporting the exit status or signal number that killed the process.