Update to use AX_CXX_COMPILE_STDCXX
This is a replacement for AX_CXX_COMPILE_STDCXX_11 (which we were using)
which also supports newer C++ standards versions which will be useful.

For C++11 the only difference seems to be that the macro now checks for
attribute support - we use C++11 attributes so that seems a good thing.