Use observed rate to help estimate value ranges
This extends the approach now used for positional filter estimates
by track how many documents each value range accepts and rejects
and using that ratio to calculate a "dynamic" estimate for the
value range.

We then use a weighted average of this and the existing "static"
estimate (calculated based on the range ends and value slot bounds)
with the weighting used being the proportion of the value slot frequency
seen - i.e. (accepts + rejects) / value_freq