Stop trying to check for incompatible GCC ABI
These checks were helpful in the GCC 3 days, but ABI versions 2 and up
are compatible aside from obscure corner cases, and GCC now defaults to
using the latest ABI version it supports.  The result is that this check
is no longer useful enough to justify the noise.

We still check for incompatible _GLIBCXX_DEBUG between the library and
application builds, since that will cause things not to work, and the
error message doesn't make it clear what's wrong.