omindexcheck: Rework handling of optional terms
With libextractor the required plugin may not be installed, but
libextractor will still return generic metadata, so testcases for
libextractor formats previously had a SKIP_IF_NO_TERMS flag.

However that prevents us from testing those generic metadata terms
so instead we now have an OPT(T) macro which can be used to mark
particular terms as optional - either none or all of the terms
marked with OPT() should to be present.