[php] Redo functor reference tracking
We need to keep hold of a reference to Xapian functor objects which are
subclassed in PHP.  We were trying to use Xapian's optional reference
counting for this, but it turns out that doesn't work in the case where
the functor is set then the object it is set on is deleted, and then
the same functor object is set on another object.

It appears to be impossible to make Xapian's optional reference counting
work with SWIG director objects, so go back to keeping references via
PHP properties, but now implemented in C++ code rather than PHP code
because the classes are no longer defined in a xapian.php wrapper.

See #186.

(cherry picked from commit 7445f6c1d5b15b946ca77724895aeadabb125131)