Adjust worker API to reduce copying of data
The handler code now calls back with results rather than filling in
several std::string& parameters passed in, which means that we can
avoid copying in the typical case where we don't get data out of
the extraction library as std::string.

The number of pages is no longer passed as a string value, which
avoids having to convert it to a string when we get it as an integer.