Add explicit initialisation phase to workers
It's fairly common for libraries to need initialising before first
use, so add an explicit function which is called which can do this.
This avoids having to check if initialisation is required on every
call to extract(), and means handling of exceptions thrown during
initialisation can be done in the common code.