Fix --mimetype to update max_ext_len correctly
We keep track of the length of the longest extension we have a
mapping for, but this was being updated using the length of the
MIME type rather than the length of the extension.  Theoretically
this could have led to us effectively ignoring --mimetype option,
but in the real world the MIME type will probably always be longer so
this just results in us testing long extensions unnecessarily.

(cherry picked from commit 986bb493902b03a0bb014e538586f156a3b05550)