Factor out helper to negate an unsigned integer
Negating an unsigned integer is valid C++, but MSVC warns about it
by default.  Unlike some MSVC warnings, this one seems helpful in
general, but we legitimately do it in a few places in the code
(e.g. as a trick to find the minimum non-zero unsigned value, or
while converting to/from a string or serialising double values).

Unfortunately there doesn't seem to be a simple way to suppress this
warning with a code tweak.  Adding conditional uses of the #pragma
to suppress this warning to each call site gets messy, so I've factored
out an negate_unsigned() template function which performs the operation
with the warning suppressed.