Revert "Enable collapsekey2 for multi && remote"
While this testcase now passes on git master, it still fails on this
branch and the changes which make it pass are too extensive to backport
(they look to be due to merging remote results in a very different way).

The failure doesn't reflect a bug as such, just that we don't give quite
such good bounds in this case as we could, so the testcase can't
determine whether that the mechanism to reduce bounds when collapsing is
working.

This reverts commit 65339b7b6d76d3908c67eaf71f7240a891561733.