Really really fix _FORTIFY_SOURCE vs mingw-w64
Including a header to check if we're on mingw-w64 results in processing
the current _FORTIFY_SOURCE setting, and we can't then usefully change
it based on whether we're on mingw-w64 or not.

Just take the simpler approach of not auto-enabling _FORTIFY_SOURCE on
mingw or mingw-w64.  It doesn't work on the latter without -lssp, and
I can't find evidence it's even supported on the former anyway.  If it
is, enabling manually will still work.

Really really fixes #808.