Improve honey testharness efficiency
Fix the check for an existing database to dir_exists() rather than
file_exists() - previously this check would never succeed and we'd
pointlessly rebuild the database for a testcase every time rather than
reusing an existing cached one.

Also open the source database for compaction as a Database, which
is a little more efficient than as a WritableDatabase.

These changes reduced the time to run honey tests from 51 to 20
seconds.