[testsuite] Add ability to kill DB's remote server
This provides a way to test handling of this situation.

It's currently only supported for a single remotetcp shard.

Supporting it for remoteprog would require a way to get the PID of the
server process, which is started by the library.  We could use a wrapper
script, or add a --pidfile option to the server itself (which may be
useful for users).

Supporting it for sharded databases would require a way to get the
Database::internal pointer for each shard (or at least each remote
shard).  The testharness could grab a copy of these as it constructs
the database.  We could potentially also support a way to get the
n-th shard from a Database object in the public API (e.g. operator[]).