xapian-tcpsrv: Don't test open DBs with --one-shot
We do a test open of the specified databases in xapian-tcpsrv so
that if they can't be opened the error is shown on invocation
rather than when the first client connects.

This doesn't seem very useful in --one-shot mode, so we now skip
doing it when this option is specified.  This also reduces overhead
in the test harness.