Fix transactions on sharded databases
commit_transaction() and cancel_transaction() weren't actually hooked
up properly and always complained that a transaction wasn't in
progress.  Looks like this doesn't affect 1.4.x (probably introduced
by the reimplementation of Database and WritableDatabase in
145503bbe4a5bf702cd13cb2e592111e8d7ca89a).

Spotted while working on enabling more testcases to run on sharded
databases, and that change (coming soon) will provide a regression test
for this.