Fix metadata tests
Seems I failed to run the right tests after the last change, and
it introduced failures for test harness backends which aren't
writable.

Mark the existing tests are requiring `writable`, and add a test
with a generated database so we have some converage when test
harness backends which aren't `writable`.