Removed unused EmptyDatabase::end_transaction_()
This method should be called end_transaction(), but actually it is
not needed so just remove it.

The base class implementation will fail with InvalidOperationError
"not in a transaction" which seems totally appropriate.

We do still need to overload begin_transaction() because otherwise the
transaction would start successfully whereas it's better to fail early.