[honey] Throw for unhandled glass conversions
We don't currently handle converting a glass database which uses
a value slot > 127 (because the glass key encoding doesn't give
keys that are always in value slot numerical order).  Previously
we had an assertion for this case, but that's bogus as it's a
totally valid input so throw UnimplementedError instead.