[honey] Improve handling of large gaps in docids
When converting from glass, a large docid gap within a glass chunk would
be converted to a large gap within the honey chunk, but honey gaps are
encoded as O(gap size) so that is undesirable.

Now the glass chunk is split at every gap, then we let the existing
chunk merging decide which chunks to recombine.

Fixes testcase valueweightsource5 for honey.