Handle value slots >127 in glass->honey conversion
Glass uses unhelpful encodings for the slot number in value slot keys
(pack_uint_last) and value slot chunks (pack_uint), neither of which
in general sort in the same order as the slot numbers.  Honey fixes
this, but that means we need to reorder the entries when converting
from glass (previously we didn't reorder, but only handled slot numbers
up to 127, which happen to sort suitably for both encodings).