backport-source-packages: More robust maint handling
Only some packages have maintclean and maint targets, and previously
we allowed for this by trying to run these targets but ignoring any
errors.  However this isn't helpful if these targets exist but fail
with a genuine error, so now we check if the targets are present and
only try to run them if they are, failing if there's an error.