backport-source-packages: Drop python2
I run the script in debian/unstable which no longer has support
for python2 in the dh helpers so building the source package fails.

It should be possible to work around this, but I suspect there just
isn't demand for newer versions of the Python2 bindings packages
any more so I'm taking the easy option, and we can review if we get
requests for them.