Fix GCC warning with --enable-log
Turn an `if` check for a "impossible" condition into an assertion.
This check is only compiled with --enable-log and seems to cause
the compiler to reason that if there's a check for it, it's possible
for the value to be out of range, and it then warns for code lower
down which just assumes that the value is in range.

(cherry picked from commit 98f25ced7bda92cac011306925fc6af1c0cffee5)