Fix fallback code for old libmagic
The code we were using seems like it would never have worked as the
return value from magic_file() is const char* and it seems also was back
then.  Nobody's reported this (I noticed while looking at the code) so
we could just require libmagic >= 4.22, but it's trivial to actually
handle so let's fix the fallback code and keep it for now.