Fix mingw32 build with C++17 compiler
This ends up with std::byte conflicting with byte typedefs in system
headers, due to `using namespace std;` in some of our internal headers.
Switch the problematic ones to more targetted `using std::string;`, etc
to fix this.

This change is specific to RELEASE/1.4 as we completely cleaned up git
master to use explicit std:: qualifying in internal headers, but that
seemed too invasive to backport (also the code has changed so much that
the 1.4 version would need many additional changes).