configure: Fix problem in probe for timer_create
The "`man sigevent`" in a #error seems to actually end up running
that as a shell command, leading to an error on stderr if there's
no sigevent man page.  Presumably on NetBSD we actually try to
compile the sigevent man page, but in a case where we deliberately
trigger a #error anyway.

(cherry picked from commit e9798d66d672d0373300717f217616a62411a692)