Fix wdf_upper_bound in ChertModifiedPostList
This was using the wdf_upper_bound from the unmodified ChertPostList
which is wrong in general as it doesn't take into account any higher
wdfs in the pending changes.

Fix by just using the global wdf upper bound in this case, which is
often not a particularly tight bound, but we need a minimally invasive
fix for chert at this point.

Bug introduced in 1.4.19 by 76d757f5f95ee73ac4090e6a76c325896dc04b27
and discovered because it causes testsuite failures if you configure
with --enable-assertions.