[Tcl] improve configure probe for stub library
The probe now works if the tclConfig.sh we find forwards to a
different script, as is the case with /usr/lib/tclConfig.sh on
current Debian.  On Debian at least our method for finding the
tclConfig.sh to use doesn't find such a forwarding script, but
it seems better to be robust to this.

(cherry picked from commit 3cc95e796b58d08a76e9d86568e3cd4f6eac028f)