Eliminate remaining uses of sprintf in C++ code
We should be able to assume snprintf() is provided now, as it was
added in C++11 (as part of the C99 compatibility).  For now we continue
to aim to support older snprintf implementations with different return
value semantics.

I've backported this patch to eliminate deprecation warnings when
building on macOS, but in the backported version of the patch we still
fall back to sprintf() in a few cases if snprintf() (or _snprintf())
isn't available as we want to minimise risk of breaking the build in the
stable release branch.

(cherry picked from commit 197c15a3e8d9cf1c483857ffbf8ebcec3587d5f9)