Handle SOCKET value we can't handle differently
We rely on being able to cast SOCKET values to int, and except for
INVALID_SOCKET to get a positive result.  It seems in practice this
always works even when SOCKET is a 64 bit type, but we have a check
in case this ever changes.  We did throw a Xapian::NetworkError
exception in this case, but that makes it harder to reuse this code
in omega, so instead close the socket and return with a WSAMFILE
error code.

(cherry picked from commit bd38e31933984e11a26048ff037247fce64da2ce)