[ci] Move mingw build to GHA
GHA has the big advantage of running jobs in parallel whereas appveyor
only runs one job at a time so a complete run can take hours.  We also
now benefit from bootstrapping the tree once and on Linux so no longer
need maintainer-only dependencies.

The GHA jobs also build xapian-letor, and we now have a UCRT64 build
(which as of late 2022 is the recommended configuration to use for
mingw64).