XFAIL remotefailure testcases which fail on mingw
We really should resolve this, but it's unhelpful to have tests
failing in master and it's unclear what's going on - the affected
testcases all seem to pass with MSVC so the code to handle this in the
test harness isn't simply broken.  I also added some debug prints
on a branch which looked plausible under mingw.