Add workaround for testsuite failures under Wine
Attempting to unlink a stub file sometimes fails with
errno == EPERM and _doserrno == ERROR_SHARING_VIOLATION.  This is
what you'd get if the file was still open, but we've already closed
it and sleeping for a second and retrying makes it work.  It'd be
better to get to the bottom of what's going on, but I've run out
of ideas and this workaround is only in the testharness at least.

(cherry picked from commit 1bb21ebad5d69747ca69b4ae607e3b4d9a9b206f)