[honey] Fix bug decoding tail-ngram entries
In the unusual case where the prefix to reuse from the key in the
first term overlaps with the known two byte suffix we weren't
correctly decoding the entry (which could have knock-on effects
for decoding subsequent terms with the same two last bytes.

This encoding isn't really sensible as in most cases we can't reuse
anything from the key in the first entry and at most there are two
bytes to reuse, so it'd be more efficient to just not store the
first reuse count byte, but for now just fix the decoding to work
correctly.