[honey] Fix compaction when merging databases
There were two bugs that affect this case:

We were zeroing the tf before using it to compute the wdf value for
terms which have a flat wdf.  This only matters when compaction is
merging databases as the computed wdf is only used then.

We were failing to correctly normalise the doclen chunk keys.