Allow using UBSan's unsigned-overflow check
In C and C++ unsigned overflow has defined behaviour (unlike signed
overflow) but in many cases it happening is not intended and indicates
a bug.

UBSan supports an unsigned-overflow check (at least the clang version
does - currently the GCC version doesn't seem to) so I've annotated the
few places in the code where we deliberately rely on unsigned overflow
(and eliminated a few cases where we can easily not rely on it) and
added this check to the clang-sanitisers job.