omindex: Fix shell emulation env var corner case
We no longer set environment variables which start with a digit, which
matches shell behaviour and also the existing comment above this code.

Spotted from reading the code - in practice this isn't a case that's
likely to be encountered, and the previous behaviour doesn't appear
to have any security consequences (if an attacker is able to specify the
commands omindex runs to extract text from files then you have a problem
already).