Fix probe for rdoc
The aim is to try to use the rdoc which matches the Ruby interpreter
found as a default (unless the user has explicitly overridden by
running ./configure RDOC=/path/to/rdoc or similar), but the way that
was being done didn't actually work and this default never kicked in.

This is likely to only be a problem if you're overriding
RUBY=/path/to/ruby but not overriding RDOC as well.

Reported by Vitaly Chikunov on the xapian-devel list.

(cherry picked from commit 53d0c3a6d0f66b7476e81cbf3f0fba929a5d50e0)