Compress empty non-optional table to zero size
Previously a non-optional table would always be compacted to at least
one block. This isn't important in general as the non-optional tables
are not likely to be empty in a real database but it's helpful for
making small test database and it seems weird that compaction was making
a database much larger in percentage terms in this edge case.